Place lambdas and layers in same stack
Once in different stacks the layer stack won't be able to
update as the outputs used by the lambda stack won't allow it
to change. Could be solved by adding an SSM parameter
but as it costs more and would make the CDK stack include more
resource it is better to have it in the same stack.